requirements.txt
# Web Framework and UI Components
dash==2.17.1
dash-bootstrap-components==1.6.1
plotly==5.24.1

# LLM and AI Libraries
langchain==0.3.1
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.1.20
langchain-ollama==0.1.1
langgraph==0.1.52

# API and HTTP
openai==1.47.1
requests==2.32.3

# Data Processing and Embeddings
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.8.2

# Text Processing
beautifulsoup4==4.12.3
argostranslate==1.9.3

# Environment and Configuration
python-dotenv==1.0.1

# Data Formats
jsonschema==4.23.0

# Optional: For development
# pytest==7.4.4
# black==24.1.1
# flake8==7.1.1
